chore(cpn): move model edit setup classes to own code files
